Fixes for artifacts writer; need separate writers for feature types
Needed to separate the rolled vs unrolled feature writer; also
everything has to maintain its own iteration counts internally.  This is
somewhat sketchy because to do this it has to be assumed that the
FeatureRun's internals will always return things in the exact same order
everytime.  That appears to be the case, but who knows if that stays
that way?  It seems like it should because the order of features defined
in a method matters just as much as the order of params in an interation
of a rolled up feature.  Bottom line, the order of these things
definitely matters so relying on that from the internals, though not
explicitly documented, seems like a reasonable assumption.